+++ src/service/Implement/CircleFriendVideoServiceImpl.ts
@@ -2,7 +2,16 @@ import CircleFriendVideo from '../../model/CircleFriendVideo';
 import CircleFriendVideoService from '../CircleFriendVideoService';
 
 class CircleFriendVideoServiceImpl implements CircleFriendVideoService {
+	async findVideoByCircleFriendId(id: number): Promise<boolean> {
+		const res = await CircleFriendVideo.findOne({
+			where: { circleFriendId: id }
+		});
+		return res ? true : false;
+	}
+
 	async removeVideoByCircleFriendId(id: number): Promise<boolean> {
+		const res1 = await this.findVideoByCircleFriendId(id);
+		if (!res1) return true;
 		const res = await CircleFriendVideo.destroy({
 			where: { circleFriendId: id }
 		});
